Rename ConsoleCustomNetworkFilter
